only one file for virtual transactions
