+++ backend/src/graphql/resolver/TransactionResolver.ts
@@ -30,8 +30,7 @@ import { calculateBalance, isHexPublicKey } from '@/util/validate'
 import { RIGHTS } from '@/auth/RIGHTS'
 import { User } from '@model/User'
 import { communityUser } from '@/util/communityUser'
-import { virtualLinkTransaction } from '@/util/virtualLinkTransaction'
-import { virtualDecayTransaction } from '@/util/virtualDecayTransaction'
+import { virtualLinkTransaction, virtualDecayTransaction } from '@/util/virtualTransactions'
 import Decimal from 'decimal.js-light'
 import { calculateDecay } from '@/util/decay'
